Create draft workflow and update link in jekyll config

+++ .Rprofile
@@ -12,51 +12,54 @@ new_draft <- function(draftname = "new_draft") {
   }
 }
 
-gen_blog <- function(input = c(".", list.dirs("_source")), output =
-                     c(".", rep("rmd_posts/_posts", length(list.dirs("_source")))),
-                     command = "bundle exec jekyll build", ...) {
-  servr::jekyll(input = input, output = output, serve = FALSE,
-                command = command, ...)
+gen_workshops <- function(
+    input = c(".", list.dirs("_source")),
+    output = c(".", rep("rmd_workshops/_posts", length(list.dirs("_source")))),
+    command = "bundle exec jekyll build", ...) {
+      dir.create(file.path("rmd_workshops"), showWarnings = FALSE)
+      dir.create(file.path("rmd_workshops", "_posts"), showWarnings = FALSE)
+      servr::jekyll(input = input, output = output, serve = FALSE,
+                    command = command, ...)
 }
 
-serve_blog <- function (input = c(".", list.dirs("_source")), output = c(".",
-    rep("rmd_posts/_posts", length(list.dirs("_source")))),
-    command = "bundle exec jekyll build", ...)
-{
-    servr::jekyll(input = input, output = output, serve = TRUE,
-                  command = command, ...)
+serve_site <- function (
+    input = c(".", list.dirs("_source")),
+    output = c(".", rep("rmd_workshops/_posts", length(list.dirs("_source")))),
+    command = "bundle exec jekyll build", ...) {
+      dir.create(file.path("rmd_workshops"), showWarnings = FALSE)
+      dir.create(file.path("rmd_workshops", "_posts"), showWarnings = FALSE)
+      servr::jekyll(input = input, output = output, serve = TRUE,
+                    command = command, script = "build.R", ...)
 }
 
-gen_drafts <- function(input = c(".", list.dirs("_drafts")), output =
-                     c(".", rep("rmd_drafts/_posts", length(list.dirs("_drafts")))),
-                     command = "bundle exec jekyll build", ...) {
-  servr::jekyll(input = input, output = output, serve = FALSE,
-                command = command, ...)
+gen_drafts <- function(
+    input = c(".", list.dirs("_drafts")),
+    output = c(".", rep("rmd_drafts/_posts", length(list.dirs("_drafts")))),
+    command = "bundle exec jekyll build", ...) {
+      dir.create(file.path("rmd_drafts"), showWarnings = FALSE)
+      dir.create(file.path("rmd_drafts", "_posts"), showWarnings = FALSE)
+      servr::jekyll(input = input, output = output, serve = FALSE,
+                    command = command, ...)
 }
 
-serve_drafts <- function (input = c(".", list.dirs("_drafts")), output = c(".",
-    rep("rmd_drafts/_posts", length(list.dirs("_drafts")))),
-    command = "bundle exec jekyll build", ...)
-{
-    servr::jekyll(input = input, output = output, serve = TRUE,
-                  command = command, ...)
-}
-
-clean_site <- function() {
+clean_website <- function() {
   unlink("_site", recursive=TRUE)
 }
-
 
 clean_drafts <- function() {
   files <- list.files("rmd_drafts/_posts/")
   files <- files[grep("md", files)]
-  file.remove(paste("rmd_drafts/_posts/", files, sep = ""))
+  if(length(files)>1) {
+    file.remove(paste("rmd_drafts/_posts/", files, sep = ""))
+  }
 }
 
-clean_posts <- function() {
-  files <- list.files("rmd_posts/_posts/")
+clean_workshops <- function() {
+  files <- list.files("rmd_workshops/_posts/")
   files <- files[grep("md", files)]
-  file.remove(paste("rmd_posts/_posts/", files, sep = ""))
+  if(length(files) > 1) {
+    file.remove(paste("rmd_workshops/_posts/", files, sep = ""))
+  }
 }
 
 use_package <- function(p) {
@@ -65,12 +68,10 @@ use_package <- function(p) {
 library(p, character.only = TRUE)
 }
 
-
 r <- getOption("repos")
 r["CRAN"] <- "http://cran.wustl.edu/"
 options(repos = r)
 rm(r)
-
 
 #use_package('TDA')
 #use_package('servr')
